Refactor(database): reorganize files.
