+++ api_nestbook/app/database/db_config.py
@@ -1,0 +1,9 @@
+from app.config import settings
+
+db_config = {
+    "host": settings.MARIADB_HOST,
+    "port": settings.MARIADB_PORT,
+    "user": settings.MARIADB_USER,
+    "password": settings.MARIADB_PASSWORD,
+    "database": settings.MARIADB_DATABASE,
+}
